# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Flame Sales & Expense Tracker API",
    "version": "1.0.0",
    "description": "API for managing organizations, projects, expenses, sales, and related resources in the Flame Sales & Expense Tracker."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://flame-expense-tracker.bagumajonah3.workers.dev",
      "description": "Production server (Cloudflare Workers)"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "Flame API key in the format flame_ak_<prefix>_<secret>. Include it as an X-Api-Key header when calling these endpoints."
      }
    },
...
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/expense-categories": {
      "get": {
        "operationId": "listExpenseCategories",
        "tags": [
          "Expense Categories"
        ],
        "summary": "List expense categories",
        "description": "Returns expense categories for the current organization (or global presets during setup).",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "projectId",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Expense categories fetched successfully."
          },
          "401": {
            "description": "API key required."
          }
        }
      },
      "post": {
        "operationId": "createExpenseCategory",
        "tags": [
          "Expense Categories"
        ],
        "summary": "Create a new expense category",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "category_name": {
                    "type": "string"
                  },
                  "project_category_id": {
                    "type": "integer",
                    "nullable": true
                  },
                  "description": {
                    "type": "string",
                    "nullable": true
                  },
                  "organization_id": {
                    "type": "integer",
                    "nullable": true
                  },
                  "project_id": {
                    "type": "integer",
                    "nullable": true
                  }
                },
                "required": [
                  "category_name"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Expense category created successfully."
          },
          "401": {
            "description": "API key required."
          }
        }
      },
      "put": {
        "operationId": "updateExpenseCategory",
        "tags": [
          "Expense Categories"
        ],
        "summary": "Update an existing expense category",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer"
                  },
                  "category_name": {
                    "type": "string"
                  },
                  "project_category_id": {
                    "type": "integer",
                    "nullable": true
                  },
                  "description": {
                    "type": "string",
                    "nullable": true
                  }
                },
                "required": [
                  "id",
                  "category_name"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Expense category updated successfully."
          },
          "401": {
            "description": "API key required."
          }
        }
      },
      "delete": {
        "operationId": "deleteExpenseCategory",
        "tags": [
          "Expense Categories"
        ],
        "summary": "Delete an expense category",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Expense category deleted successfully."
          },
          "401": {
            "description": "API key required."
          }
        }
      }
    },
...
    "/api/v1/payment-methods": {
      "get": {
        "operationId": "listPaymentMethods",
        "tags": [
          "Payment Methods"
        ],
        "summary": "List payment methods",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Payment methods fetched successfully."
          },
          "401": {
            "description": "API key required."
          }
        }
      },
      "post": {
        "operationId": "createPaymentMethod",
        "tags": [
          "Payment Methods"
        ],
        "summary": "Create a new payment method",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "method_name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string",
                    "nullable": true
                  }
                },
                "required": [
                  "method_name"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment method created successfully."
          }
        }
      },
      "put": {
        "operationId": "updatePaymentMethod",
        "tags": [
          "Payment Methods"
        ],
        "summary": "Update an existing payment method",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer"
                  },
                  "method_name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string",
                    "nullable": true
                  }
                },
                "required": [
                  "id",
                  "method_name"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment method updated successfully."
          }
        }
      },
      "delete": {
        "operationId": "deletePaymentMethod",
        "tags": [
          "Payment Methods"
        ],
        "summary": "Delete a payment method",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment method deleted successfully."
          }
        }
      }
    },
...
  }
}